Crate build_re_types

This binary runs the codegen manually.

It is easiest to call this using pixi run codegen, which will set up the necessary tools.

Macros

join 🔒

This uses [rayon::scope] to spawn all closures as tasks running in parallel. It blocks until all tasks are done.

Constants

CPP_OUTPUT_DIR_PATH 🔒

DEFINITIONS_DIR_PATH 🔒

DOCS_CONTENT_DIR_PATH 🔒

ENTRYPOINT_PATH 🔒

PYTHON_OUTPUT_DIR_PATH 🔒

PYTHON_TESTING_OUTPUT_DIR_PATH 🔒

RE_TYPES_SOURCE_HASH_PATH 🔒

SNIPPETS_DIR_PATH 🔒

SNIPPETS_REF_DIR_PATH 🔒

Functions

main 🔒